bkiam: add metrics token auth
