+++ pkg/api/basic/router.go
@@ -21,8 +21,8 @@ import (
 
 	_ "iam/docs" // docs is generated by Swag CLI, you have to import it.
 	"iam/pkg/api/basic/handler"
-	"iam/pkg/api/common"
 	"iam/pkg/config"
+	"iam/pkg/middleware"
 )
 
 // Register ...
@@ -34,7 +34,7 @@ func Register(cfg *config.Config, router *gin.Engine) {
 
 	// metrics
 	metricRouter := router.Group("/metrics")
-	metricRouter.Use(common.TokenAuth(cfg.MetricToken))
+	metricRouter.Use(middleware.TokenAuth(cfg.AuthToken))
 	metricRouter.GET("", gin.WrapH(promhttp.Handler()))
 
 	// pprof
